+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ### Added
 
+- Added `spec.Pod.ClusterDomain` to explicitly set the cluster domain used for the etcd member URLs. [#2082](https://github.com/coreos/etcd-operator/pull/2082)
+
 ### Changed
 
 ### Removed
 
 ### Fixed
+
+- Don't expose unready nodes via client service. [#2063](https://github.com/coreos/etcd-operator/pull/2063)
+- Azure blob storage: use correct list prefix [#2071](https://github.com/coreos/etcd-operator/pull/2071)
 
 ### Deprecated